Bug fix Issue 238 - Exception when using a custom fuzz file

+++ src/org/zaproxy/zap/extension/fuzz/FuzzerThread.java
@@ -125,7 +125,7 @@ public class FuzzerThread implements Runnable {
 		
 		int total = 0;
 		for (FileFuzzer fuzzer : customFuzzers) {
-			total += (int)fuzzer.getLength();
+			total += fuzzer.getLength();
 		}
 		extension.scanProgress(0, total);
